Updated class hierarchy and file structure.
