+++ src/main/java/ICompletable.java
@@ -1,0 +1,5 @@
+public interface ICompletable {
+
+    public void setCompleted(boolean value);
+    
+}
